+++ Client/client_main.go
@@ -1,0 +1,28 @@
+// ProjectRoot/Client/client_main.go
+package main
+
+import (
+	"fyne.io/fyne/v2"
+	"fyne.io/fyne/v2/app"
+	"fyne.io/fyne/v2/container"
+	"fyne.io/fyne/v2/widget"
+)
+
+func main() {
+	myApp := app.New()
+
+	client := &Client{myApp: myApp}
+
+	welcomeWindow := myApp.NewWindow("Preguntados")
+	welcomeWindow.SetContent(container.NewVBox(
+		widget.NewLabel("Bienvenido a Preguntados"),
+		widget.NewButton("Iniciar sesion", func() {
+			client.showLoginWindow()
+			welcomeWindow.Hide()
+		}),
+	))
+	welcomeWindow.Resize(fyne.NewSize(400, 400))
+	welcomeWindow.Show()
+
+	myApp.Run()
+}
