+++ Server/server.go
@@ -1,3 +1,4 @@
+// ProjectRoot/Server/server.go
 package main
 
 import (
@@ -165,7 +166,6 @@ func handleConnection(conn net.Conn) {
 			}
 		}
 	}
-
 }
 
 func sendQuestionToClient(conn net.Conn) {
@@ -210,26 +210,4 @@ func checkAnswer(answer string) bool {
 	fmt.Println("Answer: ", answer)
 	fmt.Println("Current answer: ", currentAnswer)
 	return answer == currentAnswer // Simplificado para este ejemplo, deberías verificar con la respuesta correcta
-}
-
-func main() {
-	loadQuestions()
-	ln, err := net.Listen("tcp", ":8080")
-	if err != nil {
-		fmt.Println("Error al iniciar el servidor:", err)
-		return
-	}
-	defer ln.Close()
-
-	fmt.Println("Servidor escuchando en el puerto 8080...")
-
-	for {
-		conn, err := ln.Accept()
-		if err != nil {
-			fmt.Println("Error al aceptar la conexion:", err)
-			continue
-		}
-		clients[conn] = true
-		go handleConnection(conn)
-	}
 }
